style(Inputer): rename and make static final string
